refactor: move balance checking to base
